Resolve Coding Standard
